refactoring: rename modules -> pkg
