+++ models/package.go
@@ -24,8 +24,8 @@ import (
 
 	"github.com/Unknwon/com"
 
-	"github.com/Unknwon/gowalker/modules/base"
-	"github.com/Unknwon/gowalker/modules/setting"
+	"github.com/Unknwon/gowalker/pkg/base"
+	"github.com/Unknwon/gowalker/pkg/setting"
 )
 
 var (
