+++ pkg/doc/crawl.go
@@ -26,9 +26,9 @@ import (
 	"github.com/Unknwon/com"
 	"github.com/Unknwon/log"
 
-	"github.com/Unknwon/gowalker/modules/base"
-	"github.com/Unknwon/gowalker/modules/httplib"
-	"github.com/Unknwon/gowalker/modules/setting"
+	"github.com/Unknwon/gowalker/pkg/base"
+	"github.com/Unknwon/gowalker/pkg/httplib"
+	"github.com/Unknwon/gowalker/pkg/setting"
 )
 
 var (
