+++ pkg/doc/golang.go
@@ -23,8 +23,8 @@ import (
 	"github.com/Unknwon/com"
 
 	"github.com/Unknwon/gowalker/models"
-	"github.com/Unknwon/gowalker/modules/base"
-	"github.com/Unknwon/gowalker/modules/setting"
+	"github.com/Unknwon/gowalker/pkg/base"
+	"github.com/Unknwon/gowalker/pkg/setting"
 )
 
 var (
