+++ pkg/doc/struct.go
