+++ pkg/doc/vcs.go
@@ -31,7 +31,7 @@ import (
 	"github.com/Unknwon/com"
 
 	"github.com/Unknwon/gowalker/models"
-	"github.com/Unknwon/gowalker/modules/base"
+	"github.com/Unknwon/gowalker/pkg/base"
 )
 
 // TODO: specify with command line flag
