+++ pkg/doc/walker.go
